GP2HUB-140 Add rate limits to contentful clients
