+++ apps/gp2-server/src/handlers/webhooks/events-updated.ts
@@ -1,12 +1,6 @@
 /* istanbul ignore file */
-import { RateLimiter } from 'limiter';
 import { Handler } from 'aws-lambda';
-import {
-  getRestClient,
-  GraphQLClient,
-  MakeRequestOptions,
-  RestAdapter,
-} from '@asap-hub/contentful';
+
 import {
   getJWTCredentialsFactory,
   syncCalendarFactory,
@@ -14,17 +8,14 @@ import {
   webhookEventUpdatedHandlerFactory,
 } from '@asap-hub/server-common';
 import {
-  contentfulAccessToken,
-  contentfulEnvId,
-  contentfulManagementAccessToken,
-  contentfulSpaceId,
   googleApiCredentialsSecretId,
   googleApiToken,
   region,
 } from '../../config';
 import Events from '../../controllers/event.controller';
-import { EventContentfulDataProvider } from '../../data-providers/event.data-provider';
 import { getCalendarDataProvider } from '../../dependencies/calendar.dependency';
+import { getEventDataProvider } from '../../dependencies/event.dependency';
+import { getContentfulGraphQLClientFactory } from '../../dependencies/clients.dependency';
 import logger from '../../utils/logger';
 import { sentryWrapper } from '../../utils/sentry-wrapper';
 
@@ -33,57 +24,9 @@ const getJWTCredentials = getJWTCredentialsFactory({
   region,
 });
 
-const contentfulManagementApiRateLimiter = new RateLimiter({
-  tokensPerInterval: 3,
-  interval: 'second',
-});
+const graphQLClient = getContentfulGraphQLClientFactory();
 
-const contentDeliveryApiRateLimiter = new RateLimiter({
-  tokensPerInterval: 15,
-  interval: 'second',
-});
-
-class RateLimitedRestAdapter extends RestAdapter {
-  async makeRequest<R>(options: MakeRequestOptions): Promise<R> {
-    await contentfulManagementApiRateLimiter.removeTokens(1);
-    return super.makeRequest(options);
-  }
-}
-
-class RateLimitedGraphqlClient extends GraphQLClient {
-  request = (async (
-    ...args: Parameters<GraphQLClient['request']>
-  ): Promise<ReturnType<GraphQLClient['request']>> => {
-    await contentDeliveryApiRateLimiter.removeTokens(1);
-    return super.request(...args);
-  }) as GraphQLClient['request'];
-}
-
-const contentfulGraphQLClient = new RateLimitedGraphqlClient(
-  `https://graphql.contentful.com/content/v1/spaces/${contentfulSpaceId}/environments/${contentfulEnvId}`,
-  {
-    errorPolicy: 'ignore',
-    headers: {
-      authorization: `Bearer ${contentfulAccessToken}`,
-    },
-  },
-);
-
-const contentfulRestClient = getRestClient({
-  space: contentfulSpaceId,
-  accessToken: contentfulManagementAccessToken,
-  environment: contentfulEnvId,
-  apiAdapter: new RateLimitedRestAdapter({
-    accessToken: contentfulManagementAccessToken,
-  }),
-});
-
-const eventDataProvider = new EventContentfulDataProvider(
-  contentfulGraphQLClient,
-  () => contentfulRestClient,
-);
-
-const eventController = new Events(eventDataProvider);
+const eventController = new Events(getEventDataProvider(graphQLClient));
 const syncCalendar = syncCalendarFactory(
   syncEventFactory(eventController, logger),
   getJWTCredentials,
@@ -92,10 +35,7 @@ const syncCalendar = syncCalendarFactory(
 
 export const handler: Handler = sentryWrapper(
   webhookEventUpdatedHandlerFactory(
-    getCalendarDataProvider(
-      contentfulGraphQLClient,
-      () => contentfulRestClient,
-    ),
+    getCalendarDataProvider(graphQLClient),
     syncCalendar,
     logger,
     { googleApiToken },
